+++ index.js
@@ -1,3 +1,0 @@
-const { adbm } = require('./lib/adbm')
-
-module.exports = adbm