+++ jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  moduleFileExtensions: ['ts', 'js'],
+  rootDir: 'src',
+  testRegex: '.spec.ts$',
+  transform: {
+    '^.+\\.ts$': 'ts-jest',
+  },
+  coverageDirectory: '../coverage',
+  testEnvironment: 'node',
+}