+++ src/test/mock/04-ok.js
@@ -2,15 +2,14 @@ const collection = '_mocha_temp'
 const id = 'mock_three'
 
 module.exports = {
-  failVerify: true, // used in one test to mock a failed verification, not actually part of the migration object api
-  async up (db, logger) {
+  async up(db, logger) {
     logger.debug('Inserting id %s into collection %s.', id, collection)
 
     await db.collection(collection).insertOne({ id })
   },
-  async down (db, logger) {
+  async down(db, logger) {
     logger.debug('Removing id %s from collection %s.', id, collection)
 
     await db.collection(collection).removeOne({ id })
-  }
+  },
 }